vcell-cli-utils/requirements.txt
biosimulators_utils>=0.1.85
fire
pyyaml
pandas
matplotlib
seaborn
python_libsedml
python_libsbml
